+++ modules/azurerm-mysql-flexible-server/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "Azure MySQL Flexible server."
+  value       = azurerm_mysql_flexible_server.this
+}