+++ modules/dsfhub-azure-eventhub/variables.tf
@@ -68,7 +68,7 @@ variable "eventhub_namespace" {
 }
 
 variable "format" {
-  description = "The type of audit data being sent to the Event Hub. Possible values are: AzureSQL_Managed, Blob, Cosmos_Mongo, Cosmos_SQL, Cosmos_Table, Data_Explorer, Databricks_Workspace, File, Mariadb, Mysql, Postgresql, Postgresql_Flexible, Queue, Sql, Synapse, Table. Defaults to Sql."
+  description = "The type of audit data being sent to the Event Hub. Possible values are: AzureSQL_Managed, Blob, Cosmos_Mongo, Cosmos_SQL, Cosmos_Table, Data_Explorer, Databricks_Workspace, File, Mariadb, Mysql,  Mysql_Flexible, Postgresql, Postgresql_Flexible, Queue, Sql, Synapse, Table. Defaults to Sql."
   type        = string
   default     = "Sql"
   validation {
@@ -84,6 +84,7 @@ variable "format" {
         "File",
         "Mariadb",
         "Mysql",
+        "Mysql_Flexible",
         "Postgresql",
         "Postgresql_Flexible",
         "Queue",
@@ -93,7 +94,7 @@ variable "format" {
       ],
       var.format
     )
-    error_message = "Invalid format. Possible values are: AzureSQL_Managed, Blob, Cosmos_Mongo, Cosmos_SQL, Cosmos_Table, Data_Explorer, Databricks_Workspace, File, Mariadb, Mysql, Postgresql, Postgresql_Flexible, Queue, Sql, Synapse, Table."
+    error_message = "Invalid format. Possible values are: AzureSQL_Managed, Blob, Cosmos_Mongo, Cosmos_SQL, Cosmos_Table, Data_Explorer, Databricks_Workspace, File, Mariadb, Mysql, Mysql_Flexible, Postgresql, Postgresql_Flexible, Queue, Sql, Synapse, Table."
   }
 }